Charts: split AreaChart into a private/ subdir
Move the data-validation, time-axis, and SVG-overlay helpers out of the
main file. area-chart.tsx drops from 729 to 452 lines and now only holds
the component itself plus its provider/responsive/composition wrappers.

- private/time-axis.ts: getCurveType, getFormatter, guessOptimalNumTicks
  + the per-format tick functions and X_TICK_WIDTH.
- private/default-tooltip.tsx: renderDefaultTooltip + validateData.
- private/overlays.tsx: AreaChartScalesRef + HoverGlyphs + ScaleFn /
  VisibleSeriesEntry types.
- private/index.ts: re-exports for ergonomic imports.

No behavioural changes; 21/21 tests still pass and typecheck is clean.
