+++ projects/js-packages/charts/src/charts/area-chart/area-chart.tsx
@@ -1,19 +1,7 @@
-import { formatNumberCompact, formatNumber } from '@automattic/number-formatters';
-import { curveCatmullRom, curveLinear, curveMonotoneX } from '@visx/curve';
-import { scaleTime } from '@visx/scale';
-import {
-	XYChart,
-	AreaSeries,
-	AreaStack,
-	Grid,
-	Axis,
-	DataContext,
-	TooltipContext,
-} from '@visx/xychart';
+import { formatNumberCompact } from '@automattic/number-formatters';
+import { XYChart, AreaSeries, AreaStack, Grid, Axis } from '@visx/xychart';
 import { __ } from '@wordpress/i18n';
-import { Stack } from '@wordpress/ui';
 import clsx from 'clsx';
-import { differenceInHours, differenceInYears } from 'date-fns';
 import {
 	useMemo,
 	useContext,
@@ -46,284 +34,19 @@ import { SingleChartContext, type SingleChartRef } from '../private/single-chart
 import { SvgEmptyState } from '../private/svg-empty-state';
 import { withResponsive } from '../private/with-responsive';
 import styles from './area-chart.module.scss';
-import type { AreaChartProps, AreaChartTooltipDatum } from './types';
-import type { ElementStyles, GetElementStylesParams } from '../../providers';
-import type { DataPoint, DataPointDate, SeriesData, Optional } from '../../types';
-import type { CurveType } from '../line-chart/types';
+import {
+	AreaChartScalesRef,
+	HoverGlyphs,
+	getCurveType,
+	getFormatter,
+	guessOptimalNumTicks,
+	renderDefaultTooltip,
+	validateData,
+} from './private';
+import type { AreaChartProps } from './types';
+import type { DataPointDate, Optional } from '../../types';
 import type { ResponsiveConfig } from '../private/with-responsive';
 import type { TickFormatter } from '@visx/axis';
-import type { RenderTooltipParams } from '@visx/xychart/lib/components/Tooltip';
-import type { FC, ReactNode, Ref } from 'react';
-
-const X_TICK_WIDTH = 60;
-
-const getCurveType = ( type?: CurveType, smoothing?: boolean ) => {
-	if ( ! type ) {
-		return smoothing ? curveCatmullRom : curveLinear;
-	}
-
-	switch ( type ) {
-		case 'smooth':
-			return curveCatmullRom;
-		case 'monotone':
-			return curveMonotoneX;
-		case 'linear':
-			return curveLinear;
-		default:
-			return curveLinear;
-	}
-};
-
-const renderDefaultTooltip = ( params: RenderTooltipParams< DataPointDate > ) => {
-	const { tooltipData } = params;
-	const nearestDatum = tooltipData?.nearestDatum?.datum;
-	if ( ! nearestDatum ) return null;
-
-	const tooltipPoints: AreaChartTooltipDatum[] = Object.entries( tooltipData?.datumByKey || {} )
-		.map( ( [ key, { datum } ] ) => ( {
-			key,
-			value: datum.value as number,
-		} ) )
-		.sort( ( a, b ) => b.value - a.value );
-
-	return (
-		<div className={ styles[ 'area-chart__tooltip' ] }>
-			<div className={ styles[ 'area-chart__tooltip-date' ] }>
-				{ nearestDatum.date?.toLocaleDateString() }
-			</div>
-			{ tooltipPoints.map( point => (
-				<Stack
-					key={ point.key }
-					direction="row"
-					align="center"
-					justify="space-between"
-					className={ styles[ 'area-chart__tooltip-row' ] }
-				>
-					<span className={ styles[ 'area-chart__tooltip-label' ] }>{ point.key }:</span>
-					<span className={ styles[ 'area-chart__tooltip-value' ] }>
-						{ formatNumber( point.value ) }
-					</span>
-				</Stack>
-			) ) }
-		</div>
-	);
-};
-
-const formatYearTick = ( timestamp: number ) => {
-	const date = new Date( timestamp );
-	return date.toLocaleDateString( undefined, { year: 'numeric' } );
-};
-
-const formatDateTick = ( timestamp: number ) => {
-	const date = new Date( timestamp );
-	return date.toLocaleDateString( undefined, { month: 'short', day: 'numeric' } );
-};
-
-const formatHourTick = ( timestamp: number ) => {
-	const date = new Date( timestamp );
-	return date.toLocaleTimeString( undefined, { hour: 'numeric', hour12: true } );
-};
-
-const getFormatter = ( sortedData: ReturnType< typeof useChartDataTransform > ) => {
-	const minX = Math.min( ...sortedData.map( datom => datom.data.at( 0 )?.date ) );
-	const maxX = Math.max( ...sortedData.map( datom => datom.data.at( -1 )?.date ) );
-
-	const diffInHours = Math.abs( differenceInHours( maxX, minX ) );
-	if ( diffInHours <= 24 ) {
-		return formatHourTick;
-	}
-
-	const diffInYears = Math.abs( differenceInYears( maxX, minX ) );
-	if ( diffInYears <= 1 ) {
-		return formatDateTick;
-	}
-
-	return formatYearTick;
-};
-
-const guessOptimalNumTicks = (
-	data: ReturnType< typeof useChartDataTransform >,
-	chartWidth: number,
-	tickFormatter: ( timestamp: number, index?: number, values?: unknown ) => string
-) => {
-	const minX = Math.min( ...data.map( datom => datom.data.at( 0 )?.date ) );
-	const maxX = Math.max( ...data.map( datom => datom.data.at( -1 )?.date ) );
-	const xScale = scaleTime( { domain: [ minX, maxX ] } );
-
-	const upperBound = Math.min(
-		data[ 0 ]?.data.length || 3,
-		Math.ceil( chartWidth / X_TICK_WIDTH )
-	);
-	let secondBestGuess = 1;
-
-	for ( let numTicks = upperBound; numTicks > 1; --numTicks ) {
-		const ticks = xScale.ticks( numTicks ).map( d => tickFormatter( d.getTime() ) );
-
-		if ( ticks.length > upperBound ) continue;
-
-		secondBestGuess = Math.max( secondBestGuess, ticks.length );
-
-		const uniqueTicks = Array.from( new Set( ticks ) );
-		if ( uniqueTicks.length === 1 ) return 1;
-
-		const hasConsecutiveDuplicate = ticks.some(
-			( tick, idx ) => idx > 0 && tick === ticks[ idx - 1 ]
-		);
-		if ( hasConsecutiveDuplicate ) continue;
-
-		return ticks.length;
-	}
-
-	return secondBestGuess;
-};
-
-const validateData = ( data: SeriesData[] ) => {
-	if ( ! data?.length ) return __( 'No data available', 'jetpack-charts' );
-
-	// Reject empty series.data — downstream tick formatters and stack layout
-	// don't tolerate it (Math.min(...[]) → Infinity → NaN cascades).
-	const hasEmptySeries = data.some( series => ! series.data?.length );
-	if ( hasEmptySeries ) return __( 'No data available', 'jetpack-charts' );
-
-	const hasInvalidData = data.some( series =>
-		series.data.some(
-			( point: DataPointDate | DataPoint ) =>
-				isNaN( point.value as number ) ||
-				point.value === null ||
-				point.value === undefined ||
-				( 'date' in point && point.date && isNaN( point.date.getTime() ) )
-		)
-	);
-
-	if ( hasInvalidData ) return __( 'Invalid data', 'jetpack-charts' );
-	return null;
-};
-
-const AreaChartScalesRef: FC< {
-	chartRef?: Ref< SingleChartRef >;
-	width: number;
-	height: number;
-	margin?: { top?: number; right?: number; bottom?: number; left?: number };
-} > = ( { chartRef, width, height, margin } ) => {
-	const context = useContext( DataContext );
-
-	useImperativeHandle(
-		chartRef,
-		() => ( {
-			getScales: () => {
-				if ( ! context?.xScale || ! context?.yScale ) return null;
-				return { xScale: context.xScale, yScale: context.yScale };
-			},
-			getChartDimensions: () => ( { width, height, margin: margin || {} } ),
-		} ),
-		[ context, width, height, margin ]
-	);
-
-	return null;
-};
-
-type VisibleSeriesEntry = { series: SeriesData; index: number; isVisible: boolean };
-
-// Narrowed call signature for the visx `xScale` / `yScale` we read from
-// `DataContext`. The real type is `AxisScale` (a wider d3-scale union with a
-// `NumberLike` return), but the only scales this chart configures are time +
-// linear and the call site wraps the result in `Number(...)` + `isFinite`,
-// so this narrower shape is safe in practice without spreading `any`.
-type ScaleFn = ( input: Date | number ) => number;
-
-// SVG overlay rendering a circle at each visible series for the currently
-// hovered x-position. visx's `showSeriesGlyphs` doesn't work for AreaStack
-// (registered yAccessor expects a stack-bar but receives the unwrapped
-// DataPointDate, returning NaN), so glyph positions are computed from the
-// chart's own scales. Stacked + offset='none' renders at the cumulative top
-// edge of each band, matching d3-stack semantics (missing values count as 0
-// in the running total). Other stack offsets ('expand', 'wiggle', 'silhouette')
-// are skipped — recovering exact positions there would require re-running the
-// d3-stack layout. Unstacked renders at the series' raw y-value.
-const HoverGlyphs: FC< {
-	visibleSeries: VisibleSeriesEntry[];
-	stacked: boolean;
-	stackOffset: 'none' | 'expand' | 'wiggle' | 'silhouette';
-	getElementStyles: ( params: GetElementStylesParams ) => ElementStyles;
-	strokeColor: string;
-} > = ( { visibleSeries, stacked, stackOffset, getElementStyles, strokeColor } ) => {
-	const dataContext = useContext( DataContext );
-	const tooltipContext = useContext( TooltipContext );
-
-	const xScale = dataContext?.xScale as ScaleFn | undefined;
-	const yScale = dataContext?.yScale as ScaleFn | undefined;
-	const tooltipOpen = tooltipContext?.tooltipOpen;
-	const nearestDatum = tooltipContext?.tooltipData?.nearestDatum?.datum as
-		| DataPointDate
-		| undefined;
-
-	if (
-		! tooltipOpen ||
-		! xScale ||
-		! yScale ||
-		! nearestDatum ||
-		! nearestDatum.date ||
-		( stacked && stackOffset !== 'none' )
-	) {
-		return null;
-	}
-
-	const xPx = Number( xScale( nearestDatum.date ) );
-	if ( ! Number.isFinite( xPx ) ) return null;
-
-	const hoveredTime = nearestDatum.date.getTime();
-	let cumulative = 0;
-	const circles: ReactNode[] = [];
-
-	// Iterate ALL visible series — never short-circuit — so missing-x-domain
-	// gaps don't break the cumulative offset for subsequent series.
-	for ( const { series, index } of visibleSeries ) {
-		const datum = series.data.find(
-			d => ( d as DataPointDate ).date?.getTime() === hoveredTime
-		) as DataPointDate | undefined;
-		// d3-stack treats missing-or-null values as 0 in the running total,
-		// so we coerce here and always advance the cumulative baseline before
-		// the skip guard. Without this, glyphs for series above a null point
-		// would land at the wrong stacked y.
-		const value = datum?.value ?? 0;
-		if ( stacked ) {
-			cumulative += value;
-		}
-
-		// Skip rendering a glyph when the datum is missing or null-valued —
-		// the area itself collapses to baseline at that x, so a glyph would
-		// be misleading.
-		if ( ! datum || datum.value == null ) {
-			continue;
-		}
-
-		const yPx = Number( yScale( stacked ? cumulative : value ) );
-		if ( ! Number.isFinite( yPx ) ) continue;
-
-		const { color } = getElementStyles( { data: series, index } );
-		circles.push(
-			<circle
-				key={ series.label || index }
-				cx={ xPx }
-				cy={ yPx }
-				r={ 4 }
-				fill={ color }
-				stroke={ strokeColor }
-				strokeWidth={ 1.5 }
-				paintOrder="fill"
-				data-testid={ `area-chart-hover-glyph-${ index }` }
-			/>
-		);
-	}
-
-	if ( circles.length === 0 ) return null;
-
-	return (
-		<g pointerEvents="none" className="area-chart__hover-glyphs">
-			{ circles }
-		</g>
-	);
-};
 
 const AreaChartInternal = forwardRef< SingleChartRef, AreaChartProps >(
 	(
