+++ projects/js-packages/charts/src/charts/area-chart/private/index.ts
@@ -1,0 +1,3 @@
+export { X_TICK_WIDTH, getCurveType, getFormatter, guessOptimalNumTicks } from './time-axis';
+export { renderDefaultTooltip, validateData } from './default-tooltip';
+export { AreaChartScalesRef, HoverGlyphs, type VisibleSeriesEntry } from './overlays';
